device/pnp: use correct width type for pnp_info.function
Reviewed-on: https://review.coreboot.org/c/coreboot/+/37740
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/include/device/pnp.h
@@ -35,7 +35,7 @@ extern struct device_operations pnp_ops;
 struct pnp_info {
 	struct device_operations *ops; /* LDN-specific ops override */
 #define PNP_SKIP_FUNCTION 0xffff
-	unsigned int function; /* Must be at least 16 bits (virtual LDNs)! */
+	u16 function; /* Must be at least 16 bits (virtual LDNs)! */
 	unsigned int flags;
 #define PNP_IO0  0x000001
 #define PNP_IO1  0x000002
